Remove all traces of tuplestore_donestoring() from the extension
This routine is a no-op since dd04e95 from 2003, with a macro kept
around for compatibility purposes.  This has led to the same code
patterns being copy-pasted around for no effect.

Compilation is failing in PG 17 as it has been removed so need to remove
all the reference.

Reported by: sandeep-edb