fix possible buffer overflows
When reading a string from scanf functions, it is important to always
provide a maximum width for the field.

The idiomatic way in postgres is to allocate one size larger than
MAXPGPATH and to use it in the format.

Extra care should be taken to zero the full size, as well as to copy to
other regions that may be smaller.

This was identified by SonarQube.